+++ app/Http/Controllers/MemberImgController.php
@@ -2,28 +2,19 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
 // Import Class MemberImgRequest
 use App\Http\Requests\MemberImgRequest;
-// Import DB members
-use App\Models\Member;
-// Import DB membersImg
-use App\Models\MemberImg;
+// Import DB membersImg DB members
+use App\Models\{Member, MemberImg};
 use Illuminate\Support\Facades\Storage;
 
 class MemberImgController extends Controller
 {
-
     // CREATE
     public function create($slug)
     {
-
         $member = Member::with('membersImgs')->where('slug', $slug)->first();
-
-        if (!$member->isOwner()) {
-            return redirect('/member-img/' . $member->slug)->with('msg', 'Anda tidak memiliki akses');
-        }
-
+        $this->authorize('isOwner', $member);
         return view('dashboard_create.member_img_create', compact('member'));
     }
 
@@ -57,9 +48,7 @@ class MemberImgController extends Controller
     public function edit($id)
     {
         $memberImg = MemberImg::findOrFail($id);
-        if (!$memberImg->isOwner()) {
-            return redirect('/member-img/' . $memberImg->member->slug)->with('msg', 'Anda tidak memiliki akses');
-        }
+        $this->authorize('isOwner', $memberImg);
         return view('dashboard_edit.member_img_edit', compact('memberImg'));
     }
 
@@ -67,21 +56,22 @@ class MemberImgController extends Controller
     public function update(MemberImgRequest $request, $id)
     {
         $data = $request->all();
-        $model = MemberImg::findOrFail($id)->first();
+        $model = MemberImg::findOrFail($id);
         // Img Store
         if ($request->has('img')) {
             Storage::delete($model->img);
             $img = $request->file('img');
             $data['img'] =  $request->file('img')->storeAs('img_members', time() . '.' . $img->getClientOriginalExtension());
         }
-        MemberImg::findOrFail($id)->update($data);
+        $model->update($data);
         return redirect('/member-img/' . $model->member->slug)->with('msg', 'Data Galeri Anggota Berhasil diedit');
     }
 
     // DELETE
     public function destroy($id)
     {
-        $model = MemberImg::findOrFail($id)->first();
+        $model = MemberImg::findOrFail($id);
+        $this->authorize('isOwner', $model);
         Storage::delete($model->img);
         MemberImg::destroy($id);
         return redirect('/member-img/' . $model->member->slug)->with('msg', 'Data Galeri Anggota Berhasil dihapus');
